split off PopulationInterface
