+++ docs/source/api_reference/framework/population/interface.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.population.interface
